requirements.txt
# core
numpy
pillow
matplotlib
opencv-python
scikit-learn

# jupyter (optional - for using the notebook)
jupyter

# streamlit for demo app
streamlit

# torch - install method below
torch
torchvision

# utilities
tqdm
